OUTER WALLS
Adds outer walls as a separate class holding an arraylist of walls, initialized to border the screen.
Adds a draw method to that class.
Adds a collision method to that class.
Adds a dispose method to that class.
Adds a dispose method to the Wall class.

Changes apples to always randomize within the bounds of the walls, not on them.
Changes order of collision detection so wall collision does not immediately set game over to false after a body collision.