WALLS (#1)
* WALLS

Adds outer walls as a game object and a sprite that displays on the screen with an initialization method to put them properly

* WALLS

Adds collision to the outer walls.

TODO:
- Stop apples spawning on the wall or the snake
- Refactor wall groups into own classes

* OUTER WALLS

Adds outer walls as a separate class holding an arraylist of walls, initialized to border the screen.
Adds a draw method to that class.
Adds a collision method to that class.
Adds a dispose method to that class.
Adds a dispose method to the Wall class.

Changes apples to always randomize within the bounds of the walls, not on them.
Changes order of collision detection so wall collision does not immediately set game over to false after a body collision.

* DYNAMIC WALLS

Adds dynamic walls that are randomized and do not overlap with the outer walls.
Adds a check for the apple randomization that prevents it from overlapping with a dynamic wall.
Adds collision with the dynamic walls and snake head rectangle.

* WALL GAPS

Adds overloaded methods to the outer walls that allows them to be initialized with a "doorway" in any arbitrary position and of arbitrary width.