ENEMY (#2)
* ENEMY

Adds the enemy class which repeats a lot of code from the player class.

Will include a sightline and search function alongside phases. For now, just getting the enemy placed and moving on screen is the immediate goal.

* ENEMY ON SCREEN MOVING RIGHT

Adds enemy to be rendered on screen.
Adds enemy variables to Enemy class.
Adds enemy movement, only to the right and under the same movement interval as the player.

* ENEMY TURNING WITH OUTER WALLS

Adds enemy sight to scan an arbitrary distance ahead of them.
Adds enemy turning when their sight detects an outer wall in front of them. They will turn a random direction away from it (technically only 90 degrees, but finicky in that they do both at once so double back)

* WANDERING ENEMY

Adds wandering functionality to enemy to turn when it sees a wall.
Adds random turning to enemy to make its movement less predictable.

Some issues remain like when enemy turns on edge of screen.

* ENEMY COLLISION

Adds enemy collision with the player snake and game over state if player runs into enemy.

* ENEMY DEATH

Adds enemy death when it runs into the player snake, deleting it from the playing board.